Started using git agan, catching up on development since last time. Noatble changes - got climate change atlas to work (python scripts) - added country masks as well as the SREX masks - added SPECS verification routines in R - working on covariate-dependent extreme value routines: GEV, Gauss, Gumbel seem to work, GPD not yet.

+++ upload_mask.cgi
@@ -11,10 +11,10 @@ if [ "$EMAIL" = "someone@somewhere" ]; then
   . ./myvinkfoot.cgi
 fi
 
-if [ "$FORM_set" = SREX -o "$FORM_set" = eu_rivers -o "$FORM_set" = ar5_atlas ]; then
+if [ "$FORM_set" = SREX -o "$FORM_set" = eu_rivers -o "$FORM_set" = ar5_atlas -o "$FORM_set" = countries ]; then
 	for file in $FORM_set/*.txt
 	do
-		if [ -s $file -a ${file%kaal.txt} = $file ]; then
+		if [ -s $file -a ${file%kaal.txt} = $file -a ${file%nan.txt} = $file ]; then
 			cp $file data/
 			f=`basename $file .txt`
 			maskfile=data/$f.txt
